@@ -1,9 +1,9 @@
 module github.com/milvus-io/milvus-operator
 
-go 1.18
+go 1.19
 
 require (
-	cloud.google.com/go/storage v1.6.0
+	cloud.google.com/go/storage v1.10.0
 	github.com/Azure/azure-sdk-for-go/sdk/azidentity v1.3.0
 	github.com/Azure/azure-sdk-for-go/sdk/storage/azblob v1.1.0
 	github.com/Masterminds/sprig v2.22.0+incompatible
