@@ -4,8 +4,8 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
+	"go.uber.org/mock/gomock"
 	corev1 "k8s.io/api/core/v1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -53,7 +53,7 @@ func TestMilvusReconciler_ReconcilePVCs(t *testing.T) {
 
 	t.Run("sync:no_update", func(t *testing.T) {
 		defer env.Ctrl.Finish()
-		mockClient.EXPECT().Get(ctx, gomock.Any(), gomock.Any()).Do(func(_, _, obj interface{}) {
+		mockClient.EXPECT().Get(ctx, gomock.Any(), gomock.Any()).Do(func(_, _, obj interface{}, opts ...any) {
 			pvc := obj.(*corev1.PersistentVolumeClaim)
 			r.syncPVC(ctx, m.Spec.Dep.RocksMQ.Persistence.PersistentVolumeClaim, pvc)
 		}).Return(nil)
