@@ -16,9 +16,23 @@ type K8sClient interface {
 	client.Client
 }
 
+// K8sStatusClient for mock K8sClient.Status()
+type K8sStatusClient interface {
+	client.SubResourceWriter
+}
+
 // Logger for mock
 type Logger interface {
-	logr.Logger
+	Enabled() bool
+	Error(err error, msg string, keysAndValues ...interface{})
+	GetSink() logr.LogSink
+	Info(msg string, keysAndValues ...interface{})
+	V(level int) logr.Logger
+	WithCallDepth(depth int) logr.Logger
+	WithCallStackHelper() (func(), logr.Logger)
+	WithName(name string) logr.Logger
+	WithSink(sink logr.LogSink) logr.Logger
+	WithValues(keysAndValues ...interface{}) logr.Logger
 }
 
 // PulsarClient for mock
